Move auth models into separate packages
