--- modules/user/server-scala/src/main/scala/model/oauth/UserAuth.scala
@@ -1,3 +1,0 @@
-package model.oauth
-
-case class UserAuth(google: Option[GoogleAuth] = None)
