+++ modules/user/server-scala/src/main/scala/model/oauth/google/GoogleAuthTable.scala
@@ -1,18 +1,18 @@
-package model.oauth
+package model.oauth.google
 
 import com.byteslounge.slickrepo.meta.Keyed
 import model.UserTable.UserTable
 import slick.jdbc.JdbcProfile
 
-object FacebookAuthTable extends JdbcProfile {
+object GoogleAuthTable extends JdbcProfile {
 
   import api.{Table => SlickTable, _}
 
-  val name = "FACEBOOK_AUTH"
+  val name = "GOOGLE_AUTH"
 
-  class FacebookAuthTable(tag: Tag) extends SlickTable[FacebookAuth](tag, name) with Keyed[String] {
+  class GoogleAuthTable(tag: Tag) extends SlickTable[GoogleAuth](tag, name) with Keyed[String] {
 
-    def id = column[String]("FACEBOOK_ID", O.Unique)
+    def id = column[String]("GOOGLE_ID", O.Unique)
 
     def displayName = column[String]("DISPLAY_NAME")
 
@@ -20,6 +20,6 @@ object FacebookAuthTable extends JdbcProfile {
 
     def userFk = foreignKey("USER_ID_FK", userId, TableQuery[UserTable])(_.id, onUpdate=ForeignKeyAction.Restrict, onDelete=ForeignKeyAction.Cascade)
 
-    def * = (id.?, displayName, userId).mapTo[FacebookAuth]
+    def * = (id.?, displayName, userId).mapTo[GoogleAuth]
   }
 }
