+++ modules/user/server-scala/src/main/scala/repositories/auth/FacebookAuthRepository.scala
@@ -2,8 +2,8 @@ package repositories.auth
 
 import com.byteslounge.slickrepo.repository.Repository
 import javax.inject.Inject
-import model.oauth.FacebookAuth
-import model.oauth.FacebookAuthTable.FacebookAuthTable
+import model.oauth.facebook.FacebookAuthTable.FacebookAuthTable
+import model.oauth.facebook.FacebookAuth
 import slick.ast.BaseTypedType
 import slick.jdbc.JdbcProfile
